Feature/review 후기 수정 로직 변경

+++ src/main/java/com/dolphin/demo/service/CommentService.java
@@ -5,6 +5,7 @@ import com.dolphin.demo.domain.CommentImage;
 import com.dolphin.demo.domain.Member;
 import com.dolphin.demo.domain.Place;
 import com.dolphin.demo.dto.request.CommentRequestDto;
+import com.dolphin.demo.dto.request.ImageRequestDto;
 import com.dolphin.demo.dto.response.CommentResponseDto;
 import com.dolphin.demo.exception.CustomException;
 import com.dolphin.demo.jwt.UserDetailsImpl;
@@ -16,6 +17,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+
+import java.awt.*;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
@@ -121,7 +124,7 @@ public class CommentService {
 
 
     // 후기 수정하기
-    public ResponseEntity<CommentResponseDto> updateComment(Long comment_id, CommentRequestDto commentRequestDto, List<MultipartFile> multipartFile, UserDetailsImpl userDetails) throws IOException {
+    public ResponseEntity<CommentResponseDto> updateComment(Long comment_id, ImageRequestDto imageRequestDto, List<MultipartFile> multipartFile, UserDetailsImpl userDetails) throws IOException {
 
         // 로그인한 회원인지 여부 검증
         Member member = memberRepository.findByUsername(userDetails.getUsername()).orElse(null);
@@ -138,14 +141,14 @@ public class CommentService {
 
         // 해당 후기의 모든 이미지 불러오기
         List<CommentImage> image = commentImageRepository.findAllByCommentId(comment_id);
-        comment.getPlace().updateStar(commentRequestDto.getStar()-comment.getStar(),0);
+        comment.getPlace().updateStar(imageRequestDto.getStar()-comment.getStar(),0);
 
 
         // 수정된 내용 저장
         CommentRequestDto updateComment = CommentRequestDto.builder()
-                .title(commentRequestDto.getTitle())
-                .content(commentRequestDto.getContent())
-                .star(commentRequestDto.getStar())
+                .title(imageRequestDto.getTitle())
+                .content(imageRequestDto.getContent())
+                .star(imageRequestDto.getStar())
                 .build();
         comment.update(updateComment);
         commentRepository.save(comment);
@@ -154,39 +157,39 @@ public class CommentService {
         List<String> imageList = new ArrayList<>();
         // 새로 등록하는 이미지가 없는 경우
         if(multipartFile == null) {
-            // 기존 이미지가 있다면 기존 Url 불러오기
-            List<CommentImage> saveImage = new ArrayList<>();
-            for (CommentImage existImage : image) {
-                String imageUrl = existImage.getImageUrl();
-                CommentImage existImages = CommentImage.builder()
-                        .comment(comment)
-                        .imageUrl(imageUrl)
-                        .build();
-                saveImage.add(existImages);
-                imageList.add(existImages.getImageUrl());
-            }
-            commentImageRepository.saveAll(saveImage);
             return ResponseEntity.ok().body(CommentResponseDto.builder()
                     .comment_id(comment.getId())
                     .place_id(comment.getPlace().getId())
                     .nickname(comment.getMember().getNickname())
                     .title(comment.getTitle())
                     .content(comment.getContent())
-                    .imageList(imageList)
                     .star(comment.getStar())
                     .createdAt(comment.getCreatedAt())
                     .modifiedAt(comment.getModifiedAt())
                     .build());
 
+
             // 새로 등록하는 이미지가 있는 경우
         } else {
-            // S3 저장소에 있는 이미지 삭제하기
             for (int i=0; i<image.size(); i++) {
-                amazonS3Service.deleteFile(image.get(i).getImageUrl().substring(image.get(i).getImageUrl().lastIndexOf("/") + 1));
+                /**
+                 * 1. image의 ImageUrl : 기존에 등록된 이미지
+                 * 2. imageRequestDto의 ExistUrlList : 기존 이미지 중, 수정할 때 삭제되지 않고 남아있는 이미지
+                 * 3. 1, 2번을 비교하여 같은 URL이 존재할 경우 imageList에 추가하고 없을 경우 S3와 DB에서 삭제
+                 */
+                for (int j = 0; j < imageRequestDto.getExistUrlList().size(); j++) {
+                    if (image.get(i).getImageUrl().equals(imageRequestDto.getExistUrlList().get(j))) {
+                        imageList.add(imageRequestDto.getExistUrlList().get(j));
+                    } else {
+                    // S3 저장소에 있는 이미지 삭제하기
+                        amazonS3Service.deleteFile(image.get(i).getImageUrl().substring(image.get(i).getImageUrl().lastIndexOf("/") + 1));
+
+                        // DB 이미지 삭제
+                        commentImageRepository.delete(image.get(i));
+                    }
+                }
             }
 
-            // DB 이미지 삭제
-            commentImageRepository.deleteAll(image);
 
             // 새로운 이미지 등록
             List<String> imageUrlList;
